mtd: spi-nor: fix dual-parallel read of sfdp
Striping causes the read command to return the SFDP of both the upper
and lower flash chip mixed together. The SFDP code only understands how
to parse a standard SFDP buffer, so it think the flash chip's SFDP are
corrupt. Instead of striping, read from only the lower chip. This is
sufficient as long as both chips have the same SFDP.

Signed-off-by: Brandon Maier <brandon.maier@rockwellcollins.com>